Se añade funcionamiento de register

+++ src/main/kotlin/ar/edu/unsam/phm/dto/PasajeroDTO.kt
@@ -22,7 +22,7 @@ class InformacionPasajeroDTO(
     val nombre:String,
     val apellido:String,
     val telefono:String,
-    val saldo:Double,
+    val saldo:Long,
     val listaAmigos:List<TarjetaAmigoDTO>
 )
